Improved calculations of connections of effects.
For effects of methods and operators the arguments of the effects and
the one of the methods/operators are compared. If the names are equal
and start with a questionmark, a equal constraint will be added
between the internal symbolic variables. Previously this was only
possible if the matching parameters have the same index. This was
extended to also add connections if the parameters are different.