Fixed marking bug.
Caused by the change in the definition of metavariables and metavalues
(all unplanned tasks are in one metavariable) we had a bug in
retracting the resolver. Sometimes the marking was not set back.
Therefore we have to use the unaryapplied constraint intead of the
marking in order to define the metavariablesFixed marking bug.