Parsing: allow symbol 'At'
'At' is a keyword of a temporal constraint.
This did not allow it to be used as a symbol because the lexer converted
it into a token.
This commit changes the tokens of unary temporal constraint such that
they start with '(Constraint', e.g., '(Constarint At'. This way 'At' can
be used as a symbol.